Feedback
I modified almost all the .js files. I have change the presenter, now, the presenter manages the user input and sends commands to the model. I have also changed the way the code manages tab events. In the function "handleMouseMove" in the "note factory" I have added a 24 to compensate some error in the note's movement. Now it works fine.